+++ src/terraform/module.ts
@@ -62,7 +62,7 @@ const parseGenericGitRepositoryModuleSource: (resourceType: string) => {
 
 // Terraform Registry
 // https://developer.hashicorp.com/terraform/language/modules/sources#terraform-registry
-const getTerraformRegistryUri: (moduleSource: string) => vscode.Uri | undefined = (moduleSource) => {
+const getTerraformRegistryUri: (moduleSource: string, config?: vscode.WorkspaceConfiguration) => vscode.Uri | undefined = (moduleSource, config) => {
   // Private Registry
   if (moduleSource.split("/")[0].includes(".")) {
     // <hostname>/<namespace>/<name>/<provider>
@@ -70,27 +70,59 @@ const getTerraformRegistryUri: (moduleSource: string) => vscode.Uri | undefined 
   }
 
   // Public Registry
-  const baseUrl = 'https://registry.terraform.io/modules';
   const [namespace, name, provider, ...submodules] = moduleSource.split("/").filter(part => part !== '');
 
   // Submodule
   // <hostname>/<namespace>/<name>/<provider>/<submodule>
   if (submodules.length > 0) {
-    const submodulePath = `/submodules/${submodules.join("/").replace("modules/", "")}`;
-    return vscode.Uri.parse(`${baseUrl}/${namespace}/${name}/${provider}/latest${submodulePath}`);
+    const submodule = submodules.join("/").replace("modules/", "");
+    const documentationRegistry = config?.get<string>('documentationRegistry') || 'registry.terraform.io';
+    
+    let subTemplateStr = "https://registry.terraform.io/modules/{namespace}/{name}/{provider}/latest/submodules/{submodule}";
+    if (documentationRegistry === 'library.tf') {
+      subTemplateStr = "https://library.tf/modules/{namespace}/{name}/{provider}/latest/submodules/{submodule}";
+    } else if (documentationRegistry === 'search.opentofu.org') {
+      subTemplateStr = "https://search.opentofu.org/module/{namespace}/{name}/{provider}/latest/submodules/{submodule}";
+    } else if (documentationRegistry === 'custom') {
+      subTemplateStr = config?.get<string>('customSubModuleDocURLTemplate') || subTemplateStr;
+    }
+    
+    const url = subTemplateStr
+      .replace(/{namespace}/g, namespace)
+      .replace(/{name}/g, name)
+      .replace(/{provider}/g, provider)
+      .replace(/{submodule}/g, submodule);
+      
+    return vscode.Uri.parse(url);
   }
 
   // Module
   // <hostname>/<namespace>/<name>
   if (namespace && name && provider) {
-    return vscode.Uri.parse(`${baseUrl}/${moduleSource}`);
+    const documentationRegistry = config?.get<string>('documentationRegistry') || 'registry.terraform.io';
+    
+    let templateStr = "https://registry.terraform.io/modules/{namespace}/{name}/{provider}";
+    if (documentationRegistry === 'library.tf') {
+      templateStr = "https://library.tf/modules/{namespace}/{name}/{provider}";
+    } else if (documentationRegistry === 'search.opentofu.org') {
+      templateStr = "https://search.opentofu.org/module/{namespace}/{name}/{provider}";
+    } else if (documentationRegistry === 'custom') {
+      templateStr = config?.get<string>('customModuleDocURLTemplate') || templateStr;
+    }
+    
+    const url = templateStr
+      .replace(/{namespace}/g, namespace)
+      .replace(/{name}/g, name)
+      .replace(/{provider}/g, provider);
+      
+    return vscode.Uri.parse(url);
   }
 
   return;
 };
 
-export const getLineMatchResultUri: (lmr: LineMatchResult) => vscode.Uri | undefined =
-  ({ moduleSource: s }) => {
+export const getLineMatchResultUri: (lmr: LineMatchResult, config?: vscode.WorkspaceConfiguration) => vscode.Uri | undefined =
+  ({ moduleSource: s }, config) => {
     switch (true) {
       case s.includes("http://") || s.includes("https://"):
         // vscode already handles this
@@ -103,6 +135,6 @@ export const getLineMatchResultUri: (lmr: LineMatchResult) => vscode.Uri | undef
         // suffix is already a url
         return;
       default:
-        return getTerraformRegistryUri(s);
+        return getTerraformRegistryUri(s, config);
     }
   };